#3: fix: Renamed GroupId and package structure to be.bencouwberghs
