+++ app/src/main/resources/META-INF/maven/archetype.xml
@@ -1,0 +1,9 @@
+<archetype>
+  <id>app</id>
+  <sources>
+    <source>src/main/java/App.java</source>
+  </sources>
+  <testSources>
+    <source>src/test/java/AppTest.java</source>
+  </testSources>
+</archetype>
