+++ model/pom.xml
@@ -2,7 +2,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.github.bencouwberghs</groupId>
+        <groupId>be.bencouwberghs</groupId>
         <artifactId>timeless_songs</artifactId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
