+++ model/src/main/java/be/bencouwberghs/timeless_songs/TimelessSongsApplication.java
@@ -1,4 +1,4 @@
-package io.github.bencouwberghs.timeless_songs;
+package be.bencouwberghs.timeless_songs;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
