+++ model/src/main/java/be/bencouwberghs/timeless_songs/model/Band.java
@@ -1,4 +1,4 @@
-package io.github.bencouwberghs.timeless_songs.model;
+package be.bencouwberghs.timeless_songs.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -11,20 +11,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Song {
+public class Band {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    @Column(nullable = false)
+    @Column(nullable = false, unique = true)
     private String name;
-
-    @ManyToOne
-    @JoinColumn(name = "band_id", nullable = true)
-    private Band band;
-
-    @Column(nullable = true)
-    private int year;
 
     @Column(nullable = true)
     private String linkWikiPage;
