+++ model/src/main/java/be/bencouwberghs/timeless_songs/model/Song.java
@@ -1,4 +1,4 @@
-package io.github.bencouwberghs.timeless_songs.model;
+package be.bencouwberghs.timeless_songs.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -11,13 +11,20 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Band {
+public class Song {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    @Column(nullable = false, unique = true)
+    @Column(nullable = false)
     private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "band_id", nullable = true)
+    private Band band;
+
+    @Column(nullable = true)
+    private int year;
 
     @Column(nullable = true)
     private String linkWikiPage;
